move into folder and scaffold adaptor
