+++ superset/assets/src/visualizations/nvd3/nvd3_vis.css
