+++ superset/assets/src/visualizations/nvd3/nvd3_vis.js
@@ -9,18 +9,17 @@ import moment from 'moment';
 import d3tip from 'd3-tip';
 import dompurify from 'dompurify';
 
-import { getColorFromScheme } from '../modules/colors';
-import AnnotationTypes, {
-  applyNativeColumns,
-} from '../modules/AnnotationTypes';
-import { customizeToolTip, d3TimeFormatPreset, d3FormatPreset, tryNumify } from '../modules/utils';
-import { formatDateVerbose } from '../modules/dates';
-import { isTruthy, TIME_SHIFT_PATTERN } from '../utils/common';
-import { t } from '../locales';
+import { getColorFromScheme } from '../../modules/colors';
+import AnnotationTypes, { applyNativeColumns } from '../../modules/AnnotationTypes';
+import { customizeToolTip, d3TimeFormatPreset, d3FormatPreset, tryNumify } from '../../modules/utils';
+import { formatDateVerbose } from '../../modules/dates';
+import { isTruthy } from '../../utils/common';
+import { t } from '../../locales';
+import { addTotalBarValues, hideTooltips, wrapTooltip, getMaxLabelSize, formatLabel } from './utils';
 
 // CSS
 import './nvd3_vis.css';
-import { VIZ_TYPES } from './';
+import { VIZ_TYPES } from '../index';
 
 const minBarWidth = 15;
 // Limit on how large axes margins can grow as the chart window is resized
@@ -42,106 +41,14 @@ const TIMESERIES_VIZ_TYPES = [
   'time_pivot',
 ];
 
-const addTotalBarValues = function (svg, chart, data, stacked, axisFormat) {
-  const format = d3.format(axisFormat || '.3s');
-  const countSeriesDisplayed = data.length;
-
-  const totalStackedValues = stacked && data.length !== 0 ?
-    data[0].values.map(function (bar, iBar) {
-      const bars = data.map(function (series) {
-        return series.values[iBar];
-      });
-      return d3.sum(bars, function (d) {
-        return d.y;
-      });
-    }) : [];
-
-  const rectsToBeLabeled = svg.selectAll('g.nv-group').filter(
-    function (d, i) {
-      if (!stacked) {
-        return true;
-      }
-      return i === countSeriesDisplayed - 1;
-    }).selectAll('rect');
-
-  const groupLabels = svg.select('g.nv-barsWrap').append('g');
-  rectsToBeLabeled.each(
-    function (d, index) {
-      const rectObj = d3.select(this);
-      if (rectObj.attr('class').includes('positive')) {
-        const transformAttr = rectObj.attr('transform');
-        const yPos = parseFloat(rectObj.attr('y'));
-        const xPos = parseFloat(rectObj.attr('x'));
-        const rectWidth = parseFloat(rectObj.attr('width'));
-        const textEls = groupLabels.append('text')
-          .attr('x', xPos) // rough position first, fine tune later
-          .attr('y', yPos - 5)
-          .text(format(stacked ? totalStackedValues[index] : d.y))
-          .attr('transform', transformAttr)
-          .attr('class', 'bar-chart-label');
-        const labelWidth = textEls.node().getBBox().width;
-        textEls.attr('x', xPos + rectWidth / 2 - labelWidth / 2); // fine tune
-      }
-    });
-};
-
-function hideTooltips() {
-  $('.nvtooltip').css({ opacity: 0 });
-}
-
-function wrapTooltip(chart, container) {
-  const tooltipLayer = chart.useInteractiveGuideline && chart.useInteractiveGuideline() ?
-    chart.interactiveLayer : chart;
-  const tooltipGeneratorFunc = tooltipLayer.tooltip.contentGenerator();
-  tooltipLayer.tooltip.contentGenerator((d) => {
-    let tooltip = `<div style="max-width: ${container.width() * 0.5}px">`;
-    tooltip += tooltipGeneratorFunc(d);
-    tooltip += '</div>';
-    return tooltip;
-  });
-}
-
-function getMaxLabelSize(container, axisClass) {
-  // axis class = .nv-y2  // second y axis on dual line chart
-  // axis class = .nv-x  // x axis on time series line chart
-  const labelEls = container.find(`.${axisClass} text`).not('.nv-axislabel');
-  const labelDimensions = labelEls.map(i => labelEls[i].getComputedTextLength() * 0.75);
-  return Math.ceil(Math.max(...labelDimensions));
-}
-
-export function formatLabel(input, verboseMap = {}) {
-  // The input for label may be a string or an array of string
-  // When using the time shift feature, the label contains a '---' in the array
-  const verboseLkp = s => verboseMap[s] || s;
-  let label;
-  if (Array.isArray(input) && input.length) {
-    const verboseLabels = input.map(l => TIME_SHIFT_PATTERN.test(l) ? l : verboseLkp(l));
-    label = verboseLabels.join(', ');
-  } else {
-    label = verboseLkp(input);
-  }
-  return label;
-}
-
-export default function nvd3Vis(slice, payload) {
+function nvd3Vis(element, props, slice, data) {
   let chart;
   let colorKey = 'key';
-  const isExplore = $('#explore-container').length === 1;
+  const isExplore = document.querySelector('#explorer-container') !== null;
 
-  let data;
-  if (payload.data) {
-    if (Array.isArray(payload.data)) {
-        data = payload.data.map(x => ({
-            ...x, key: formatLabel(x.key, slice.datasource.verbose_map),
-        }));
-    } else {
-      data = payload.data;
-    }
-  } else {
-    data = [];
-  }
+  const container = element;
+  container.innerHTML = '';
 
-  slice.container.html('');
   slice.clearError();
 
   let width = slice.width();
@@ -886,3 +793,23 @@ export default function nvd3Vis(slice, payload) {
 
   nv.addGraph(drawGraph);
 }
+
+function adaptor(slice, payload) {
+  const { formData, datasource, selector } = slice;
+
+  const element = document.querySelector(selector);
+
+  const rawData = payload.data || [];
+  const data = Array.isArray(rawData)
+    ? rawData.map(x => ({
+      ...x,
+      key: formatLabel(x.key, datasource.verbose_map),
+    }))
+    : rawData;
+
+  const props = {};
+
+  return nvd3Vis(element, props, slice, data);
+}
+
+export default adaptor;
